+++ packages/core/src/State.ts
@@ -6,7 +6,6 @@ import { getStateValue } from './stateUtils.ts';
 import { TypegenDisabled, TypegenEnabled } from './typegenTypes.ts';
 import type {
   ProvidedActor,
-  ActorRefFrom,
   AnyMachineSnapshot,
   AnyStateMachine,
   EventObject,
@@ -16,37 +15,11 @@ import type {
   StateConfig,
   StateValue,
   AnyActorRef,
-  Compute,
   EventDescriptor,
   Snapshot,
   ParameterizedObject
 } from './types.ts';
 import { flatten, matchesState } from './utils.ts';
-
-type ComputeConcreteChildren<TActor extends ProvidedActor> = {
-  [A in TActor as 'id' extends keyof A
-    ? A['id'] & string
-    : never]?: ActorRefFrom<A['logic']>;
-};
-
-type ComputeChildren<TActor extends ProvidedActor> =
-  // only proceed further if all configured `src`s are literal strings
-  string extends TActor['src']
-    ? // TODO: replace with UnknownActorRef~
-      // TODO: consider adding `| undefined` here
-      Record<string, AnyActorRef>
-    : Compute<
-        ComputeConcreteChildren<TActor> &
-          // check if all actors have IDs
-          (undefined extends TActor['id']
-            ? // if they don't we need to create an index signature containing all possible actor types
-              {
-                [id: string]: TActor extends any
-                  ? ActorRefFrom<TActor['logic']> | undefined
-                  : never;
-              }
-            : {})
-      >;
 
 export function isMachineSnapshot<
   TContext extends MachineContext,
@@ -63,7 +36,7 @@ export function isMachineSnapshot<
 interface MachineSnapshotBase<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
@@ -71,7 +44,8 @@ interface MachineSnapshotBase<
   machine: StateMachine<
     TContext,
     TEvent,
-    TActor,
+    TChildren,
+    ProvidedActor,
     ParameterizedObject,
     ParameterizedObject,
     string,
@@ -94,7 +68,7 @@ interface MachineSnapshotBase<
   /**
    * An object mapping actor names to spawned/invoked actors.
    */
-  children: ComputeChildren<TActor>;
+  children: TChildren;
 
   /**
    * Whether the current state value is a subset of the given parent state value.
@@ -108,7 +82,7 @@ interface MachineSnapshotBase<
     this: MachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -124,7 +98,7 @@ interface MachineSnapshotBase<
     this: MachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -144,7 +118,7 @@ interface MachineSnapshotBase<
     this: MachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -156,7 +130,7 @@ interface MachineSnapshotBase<
     this: MachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -167,7 +141,7 @@ interface MachineSnapshotBase<
     this: MachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -178,14 +152,14 @@ interface MachineSnapshotBase<
 interface ActiveMachineSnapshot<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
 > extends MachineSnapshotBase<
     TContext,
     TEvent,
-    TActor,
+    TChildren,
     TTag,
     TOutput,
     TResolvedTypesMeta
@@ -198,14 +172,14 @@ interface ActiveMachineSnapshot<
 interface DoneMachineSnapshot<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
 > extends MachineSnapshotBase<
     TContext,
     TEvent,
-    TActor,
+    TChildren,
     TTag,
     TOutput,
     TResolvedTypesMeta
@@ -218,14 +192,14 @@ interface DoneMachineSnapshot<
 interface ErrorMachineSnapshot<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
 > extends MachineSnapshotBase<
     TContext,
     TEvent,
-    TActor,
+    TChildren,
     TTag,
     TOutput,
     TResolvedTypesMeta
@@ -238,14 +212,14 @@ interface ErrorMachineSnapshot<
 interface StoppedMachineSnapshot<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
 > extends MachineSnapshotBase<
     TContext,
     TEvent,
-    TActor,
+    TChildren,
     TTag,
     TOutput,
     TResolvedTypesMeta
@@ -258,7 +232,7 @@ interface StoppedMachineSnapshot<
 export type MachineSnapshot<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
@@ -266,7 +240,7 @@ export type MachineSnapshot<
   | ActiveMachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -274,7 +248,7 @@ export type MachineSnapshot<
   | DoneMachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -282,7 +256,7 @@ export type MachineSnapshot<
   | ErrorMachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -290,7 +264,7 @@ export type MachineSnapshot<
   | StoppedMachineSnapshot<
       TContext,
       TEvent,
-      TActor,
+      TChildren,
       TTag,
       TOutput,
       TResolvedTypesMeta
@@ -345,18 +319,21 @@ const machineSnapshotToJSON = function toJSON(this: AnyMachineSnapshot) {
 };
 
 const machineSnapshotGetMeta = function getMeta(this: AnyMachineSnapshot) {
-  return this._nodes.reduce((acc, stateNode) => {
-    if (stateNode.meta !== undefined) {
-      acc[stateNode.id] = stateNode.meta;
-    }
-    return acc;
-  }, {} as Record<string, any>);
+  return this._nodes.reduce(
+    (acc, stateNode) => {
+      if (stateNode.meta !== undefined) {
+        acc[stateNode.id] = stateNode.meta;
+      }
+      return acc;
+    },
+    {} as Record<string, any>
+  );
 };
 
 export function createMachineSnapshot<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TResolvedTypesMeta = TypegenDisabled
 >(
@@ -365,7 +342,7 @@ export function createMachineSnapshot<
 ): MachineSnapshot<
   TContext,
   TEvent,
-  TActor,
+  TChildren,
   TTag,
   undefined,
   TResolvedTypesMeta
@@ -403,7 +380,7 @@ export function cloneMachineSnapshot<TState extends AnyMachineSnapshot>(
 export function getPersistedState<
   TContext extends MachineContext,
   TEvent extends EventObject,
-  TActor extends ProvidedActor,
+  TChildren extends Record<string, AnyActorRef | undefined>,
   TTag extends string,
   TOutput,
   TResolvedTypesMeta = TypegenDisabled
@@ -411,7 +388,7 @@ export function getPersistedState<
   state: MachineSnapshot<
     TContext,
     TEvent,
-    TActor,
+    TChildren,
     TTag,
     TOutput,
     TResolvedTypesMeta