+++ packages/core/src/setup.ts
@@ -1,21 +1,23 @@
-import {
-  ResolveTypegenMeta,
-  StateMachine,
-  TypegenDisabled,
-  createMachine
-} from '.';
+import { StateMachine } from './StateMachine';
+import { createMachine } from './createMachine';
 import { GuardPredicate } from './guards';
+import { ResolveTypegenMeta, TypegenDisabled } from './typegenTypes';
 import {
-  AnyActorLogic,
-  MachineContext,
-  AnyEventObject,
-  NonReducibleUnknown,
-  MachineConfig,
-  Values,
-  ParameterizedObject,
   ActionFunction,
+  AnyActorLogic,
+  AnyActorRef,
+  AnyEventObject,
+  Cast,
+  DelayConfig,
+  Invert,
+  IsNever,
+  MachineConfig,
+  MachineContext,
+  NonReducibleUnknown,
+  ParameterizedObject,
   SetupTypes,
-  DelayConfig
+  ToChildren,
+  Values
 } from './types';
 
 type ToParameterizedObject<
@@ -30,31 +32,48 @@ type ToParameterizedObject<
   };
 }>;
 
-type ToProvidedActor<TActors extends Record<string, AnyActorLogic>> = Values<{
-  [K in keyof TActors & string]: {
+type DefaultToAnyActors<TActors extends Record<string, AnyActorLogic>> =
+  // if `keyof TActors` is `never` then it means that both `children` and `actors` were not supplied
+  // `never` comes from the default type of the `TChildrenMap` type parameter
+  // in such a case we "replace" `TActors` with a more traditional~ constraint
+  // one that doesn't depend on `Values<TChildrenMap>`
+  IsNever<keyof TActors> extends true ? Record<string, AnyActorLogic> : TActors;
+
+// at the moment we allow extra actors - ones that are not specified by `children`
+// this could be reconsidered in the future
+type ToProvidedActor<
+  TChildrenMap extends Record<string, string>,
+  TActors extends Record<Values<TChildrenMap>, AnyActorLogic>
+> = Values<{
+  [K in keyof DefaultToAnyActors<TActors> & string]: {
     src: K;
     logic: TActors[K];
+    id: IsNever<TChildrenMap> extends true
+      ? string | undefined
+      : K extends keyof Invert<TChildrenMap>
+        ? Invert<TChildrenMap>[K]
+        : string | undefined;
   };
 }>;
 
 export function setup<
   TContext extends MachineContext,
   TEvent extends AnyEventObject, // TODO: consider using a stricter `EventObject` here
-  TActors extends Record<string, AnyActorLogic>,
+  TActors extends Record<Values<TChildrenMap>, AnyActorLogic>,
   TActions extends Record<string, ParameterizedObject['params'] | undefined>,
   TGuards extends Record<string, ParameterizedObject['params'] | undefined>,
   TDelay extends string,
   TTag extends string,
   TInput,
-  TOutput extends NonReducibleUnknown
+  TOutput extends NonReducibleUnknown,
+  TChildrenMap extends Record<string, string> = never
 >({
   actors,
   actions,
   guards,
   delays
 }: {
-  types?: SetupTypes<TContext, TEvent, TTag, TInput, TOutput>;
-
+  types?: SetupTypes<TContext, TEvent, TChildrenMap, TTag, TInput, TOutput>;
   actors?: TActors;
   actions?: {
     [K in keyof TActions]: ActionFunction<
@@ -62,7 +81,7 @@ export function setup<
       TEvent,
       TEvent,
       TActions[K],
-      ToProvidedActor<TActors>,
+      ToProvidedActor<TChildrenMap, TActors>,
       ToParameterizedObject<TActions>,
       ToParameterizedObject<TGuards>,
       TDelay
@@ -85,11 +104,11 @@ export function setup<
     >;
   };
 }): {
-  createMachine: (
-    config: MachineConfig<
+  createMachine: <
+    const TConfig extends MachineConfig<
       TContext,
       TEvent,
-      ToProvidedActor<TActors>,
+      ToProvidedActor<TChildrenMap, TActors>,
       ToParameterizedObject<TActions>,
       ToParameterizedObject<TGuards>,
       TDelay,
@@ -99,17 +118,23 @@ export function setup<
       ResolveTypegenMeta<
         TypegenDisabled,
         TEvent,
-        ToProvidedActor<TActors>,
+        ToProvidedActor<TChildrenMap, TActors>,
         ToParameterizedObject<TActions>,
         ToParameterizedObject<TGuards>,
         TDelay,
         TTag
       >
     >
+  >(
+    config: TConfig
   ) => StateMachine<
     TContext,
     TEvent,
-    ToProvidedActor<TActors>,
+    Cast<
+      ToChildren<ToProvidedActor<TChildrenMap, TActors>>,
+      Record<string, AnyActorRef | undefined>
+    >,
+    ToProvidedActor<TChildrenMap, TActors>,
     ToParameterizedObject<TActions>,
     ToParameterizedObject<TGuards>,
     TDelay,
@@ -119,7 +144,7 @@ export function setup<
     ResolveTypegenMeta<
       TypegenDisabled,
       TEvent,
-      ToProvidedActor<TActors>,
+      ToProvidedActor<TChildrenMap, TActors>,
       ToParameterizedObject<TActions>,
       ToParameterizedObject<TGuards>,
       TDelay,
@@ -129,11 +154,11 @@ export function setup<
 } {
   return {
     createMachine: (config) =>
-      createMachine(config, {
+      (createMachine as any)(config, {
         actors,
         actions,
         guards,
         delays
-      } as any)
+      })
   };
 }